Add numbers lib/test.
