+++ decrypted.txt
@@ -1,1 +1,1 @@
-prpww zzyh
+hello world
