+++ lib/decrypt.rb
@@ -1,9 +1,11 @@
 require './lib/enigma'
+
 input, output, key, date = ARGV
 
 enigma = Enigma.new
-cypher = IO.read(input)
-message = enigma.decrypt(message, key, date)
+ciphertext = IO.read(input)
+decrypted = enigma.decrypt(ciphertext, key, date)
+
 IO.write(output, decrypted[:decryption])
 
 puts "Created '#{output}' with the key #{decrypted[:key]} and date #{decrypted[:date]}"
