+++ message.txt
@@ -1,1 +1,1 @@
-Hello world
+hello world
